requirements.txt
# Basic environment for retrieval experiments
numpy
tqdm
sentence-transformers
faiss-cpu
openai
PyPDF2
